FIXES centralize Mesh mocking into setup file, reset db between tests
