+++ package.json
@@ -13,9 +13,9 @@
         "build": "tsc -p tsconfig.json",
         "integration-test": "mocha --require source-map-support/register --require make-promises-safe lib/integration-test/**/*_test.js --timeout 200000 --exit",
         "test": "yarn test:rest && yarn test:metatx && yarn test:rfqt",
-        "test:metatx": "env-cmd -f ./test/test_env mocha --require source-map-support/register --require make-promises-safe lib/test/meta_transaction_**.js --timeout 200000 --exit --bail",
-        "test:rest": "env-cmd -f ./test/test_env mocha --exclude lib/test/rfqt_test.js --exclude lib/test/meta_transaction_test.js --require source-map-support/register --require make-promises-safe lib/test/**/*_test.js --timeout 200000 --exit --bail",
-        "test:rfqt": "env-cmd -f ./test/test_env mocha --require source-map-support/register --require make-promises-safe lib/test/rfqt_**.js --timeout 200000 --exit --bail",
+        "test:metatx": "env-cmd -f ./test/test_env mocha --file lib/test/test_setup.js --require source-map-support/register --require make-promises-safe lib/test/meta_transaction_**.js --timeout 200000 --exit --bail",
+        "test:rest": "env-cmd -f ./test/test_env mocha --file lib/test/test_setup.js --exclude lib/test/rfqt_test.js --exclude lib/test/meta_transaction_test.js --require source-map-support/register --require make-promises-safe lib/test/**/*_test.js --timeout 200000 --exit --bail",
+        "test:rfqt": "env-cmd -f ./test/test_env mocha --file lib/test/test_setup.js --require source-map-support/register --require make-promises-safe lib/test/rfqt_**.js --timeout 200000 --exit --bail",
         "dev": "nodemon -r dotenv/config src/index.ts | pino-pretty",
         "dev:service:http": "nodemon -r dotenv/config src/runners/http_service_runner.ts | pino-pretty",
         "dev:service:sra_http": "nodemon -r dotenv/config src/runners/http_sra_service_runner.ts | pino-pretty",
